swagger add documentando api

+++ app/Http/Controllers/CentralidadeController.php
@@ -12,60 +12,62 @@ use OpenApi\Annotations as OA;
 
 class CentralidadeController extends Controller
 {
-    /**
- * @OA\Info(
- *    title="Swagger with Laravel",
- *    version="1.0.0",
- * )
- * @OA\SecurityScheme(
- *     type="http",
- *     securityScheme="bearerAuth",
- *     scheme="bearer",
- *     bearerFormat="JWT"
- * )
-
- */
 
 
-    /**
-* @OA\Post(
-     *     path="/api/centralidades",
-     *     summary="Cadastra uma nova centralidade",
-     *     @OA\Parameter(
-     *         name="c_desccentr",
-     *         in="query",
-     *         description="Denominação da centralidade",
-     *         required=true,
-     *         @OA\Schema(type="string")
-     *     ),
-     *     @OA\Parameter(
-     *         name="n_codicoord",
-     *         in="query",
-     *         description="id do coordenador ou da adminstração da centelidade",
-     *         required=false,
-     *         @OA\Schema(type="int")
-     *     ),
-     *     @OA\Parameter(
-     *         name="n_codiender",
-     *         in="query",
-     *         description="id do endereço",
-     *         required=true,
-     *         @OA\Schema(type="int")
-     *     ),
-     *     @OA\Response(response="201", description="Centralidade cadastrada successfully"),
-     *     @OA\Response(response="422", description="Validation errors")
-     * )
-     */
+/**
+    * @OA\Get(
+        *     tags={"/centralidades"},
+        *     path="/api/centralidades",
+        *     summary="listar centralidades",
+        *     security={{"bearerAuth": {} }},
+        *     @OA\Response(response="200", description="successfully"),
+        *     @OA\Response(response="500", description="Erro no servidor")
+        * )
+*/
     public function getAll()
     {
+        Centralidade::paginate();
         try {
-            return Centralidade::all();
+            return response()->json(['centralidades' => Centralidade::all()],200);
         } catch (\Illuminate\Database\QueryException $e) {
             return response()->json(['message' => $e->getMessage()], 500);
         }
     }
+
+/**
+    * @OA\Post(
+        *     tags={"/centralidades"},
+        *     path="/api/centralidades",
+        *     summary="Cadastrar centralidade",
+        *     security={{"bearerAuth": {} }},
+        *     @OA\RequestBody(
+        *       required=true,
+        *       @OA\JsonContent(
+        *          type="object",
+        *          @OA\Property(property="c_desccentr",type="string",default="Centralidade B",description="denominação da centralidade"),
+        *          @OA\Property(property="n_codiender",type="int",description="id do endereço da centralidade")
+        *       )
+        *     ),
+        *     
+        *     @OA\Response(response="201", description="Centralidade cadastrada successfully"),
+        *     @OA\Response(response="412", description="Erro ao validar os dados"),
+        *     @OA\Response(response="500", description="Validation errors")
+        * )
+     */
+
+
     public function create(Request $req, Response $res)
     {
+        /*
+            o que aconteice na interfafce: o usuario digita os dados do endereço
+            e os dados da centralidade.  
+        */
+        /*TODO
+            criar o endereço
+            pegar o id do endereço
+            usa-lo para criar a centralidade
+
+         */
         $isValidData = Validator::make($req->all(), [
             "c_desccentr" => 'required|string|max:50',
             "n_nblocentr" => 'integer',
@@ -74,7 +76,7 @@ class CentralidadeController extends Controller
             "n_codiadmin" => 'integer'
         ]);
         if ($isValidData->fails())
-            return response()->json(['erros' => $isValidData->errors(), 'message' => 'erro ao validar os dados'], 400);
+            return response()->json(['erros' => $isValidData->errors(), 'message' => 'erro ao validar os dados'], 412);
 
         try {
             Centralidade::create($req->all());
@@ -83,32 +85,63 @@ class CentralidadeController extends Controller
             return response()->json(['message' => $e->getMessage()], 500);
         }
     }
-    public function getAllByProvincia($denominacao){
-
-        /* select que faz isso
-          select * from tracentr, traender where tracentr.n_codiender = traender.n_codiender and traender.c_provender = 'Luanda';
-        */
-        try {
-            Centralidade::findOrFail($idCentralidade);
-            return Bloco::where('n_codicentr', '=', $idCentralidade)->get();
-        } catch (QueryException $e) {
-            return response()->json(['message' => $e->getMessage()], 500);
-        }
-    }
+    /**
+    * @OA\Delete(
+        *     tags={"/centralidades"},
+        *     path="/api/centralidades/{centralidade}",
+        *     summary="apagar uma centralidade",
+        *       security={{"bearerAuth": {} }},
+        *       @OA\Parameter(
+        *         name="centralidade",
+        *         in="path",
+        *         description="id da centralidade",
+        *         required=false,
+        *         @OA\Schema(type="int")
+        *     ),
+        *     @OA\Response(response="200", description="centralidade deletada com sucesso!"),
+        *     @OA\Response(response="404", description="centralidade não encontrada"),
+        *     @OA\Response(response="500", description="Erro no servidor")
+        * )
+     */
     public function delete($id)
     {
         try {
-            $centralidade = Centralidade::findOrFail($id);
+            $centralidade = Centralidade::find($id);
+            if(! $centralidade)
+                return response()->json(['message' => 'centralidade não encontrada'], 404);
             $centralidade->delete();
             return response()->json(['message' => "centralidade deletada com sucesso!"], 200);
         } catch (\Illuminate\Database\QueryException $e) {
             return response()->json(['message' => $e->getMessage()], 500);
         }
     }
+    /**
+    * @OA\Put(
+    *     tags={"/centralidades"},
+    *     path="/api/centralidades",
+    *     summary="Atualizar um objecto no banco de dado",
+    *     description="atualizar os dados de uma centralidade",
+    *     security={{"bearerAuth": {} }},
+    *     @OA\RequestBody(
+    *       required=true,
+    *       @OA\JsonContent(
+    *          type="object",
+    *          @OA\Property(property="c_desccentr",type="string",default="Centralidade B",description="denominação da centralidade"),
+    *          @OA\Property(property="n_codiender",type="int",description="id do endereço da centralidade")
+    *       )
+    *     ),
+    *     
+    *     @OA\Response(response="201", description="Centralidade atualizada com sucesso"),
+    *     @OA\Response(response="412", description="Erro ao validar os dados"),
+    *     @OA\Response(response="500", description="Erro no servidor")
+    * )
+     */
     public function update(Request $req, $id)
     {
         try {
-            $centralidade = Centralidade::findOrFail($id);
+            $centralidade = Centralidade::find($id);
+            if(!$centralidade)
+                return response()->json(['message' => 'centralidade não encontrada'], 404);
 
             $centralidade->update($req->all());
 
@@ -117,11 +150,31 @@ class CentralidadeController extends Controller
             return response()->json(['message' => $e->getMessage()], 500);
         }
     }
-
+    /**
+    * @OA\Get(
+        *     tags={"/centralidades"},
+        *     path="/api/centralidades/{centralidade}",
+        *     summary="mostrar uma centralidade",
+        *     security={{ "bearerAuth": {}}},   
+        *     @OA\Parameter(
+        *         name="centralidade",
+        *         in="path",
+        *         description="id da centralidade",
+        *         required=false,
+        *         @OA\Schema(type="int")
+        *     ),
+        *     @OA\Response(response="200", description="sucesso"),
+        *     @OA\Response(response="404", description="centralidade não encontrada"),
+        *     @OA\Response(response="500", description="Erro no servidor")
+        * )
+     */
     public function getOne($id)
     {
         try {
-            $centralidade = Centralidade::findOrFail($id);
+            $centralidade = Centralidade::find($id);
+            if(!$centralidade)
+                return response()->json(['message' => 'centralidade não encontrada'], 404);
+
             return response()->json($centralidade, 200);
         } catch (\Illuminate\Database\QueryException $e) {
             return response()->json(['message' => $e->getMessage()], 500);
